support_bot/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
drf-yasg==1.20.0
django-environ==0.9.0
pyTelegramBotAPI==4.12.0
requests==2.31.0
openpyxl==3.1.2
python-dotenv
pandas==1.3.3
numpy==1.21.2
xlsxwriter==3.0.9
